woops

+++ sablec/src/main.rs
@@ -16,7 +16,7 @@ use sable_parser::{
 };
 
 const SOURCE: &str = r#"
-func main(argc: i32, argv: i32**): i32 {
+func main(argc: i32, argv: i8**): i32 {
   var lol = 123;
 }
 "#;
